refactor(home): app/에 있던 파일들 전부 features layer로 변경 (#51)
