+++ src/features/home/HomeActionCards.tsx
@@ -1,7 +1,7 @@
 // 홈 액션 카드 그리드 — 배틀/도감 카드 2종 목록
 
 import { homeActionCards } from '@/app/(main)/home/_constants/home';
-import HomeActionCard from '@/app/(main)/home/_components/HomeActionCard';
+import HomeActionCard from '@/features/home/HomeActionCard';
 
 export default function HomeActionCards() {
   return (
