+++ src/features/home/TrainerStatusBar.tsx
